tty/uart16550: fixes and performance improvements
Fix FIFO threshold not being set correctly on some platforms.
Start handler thread after initializing hardware, not before.
Add compile-time options for setting FIFO threshold.
Use library implementation of software FIFO instead of dedicated buffer
implementation.
Read and accumulate line status register to detect error conditions.
Read as much data as is available in ISR.
Increase stack size which is necessary for AArch64.

YT: RTOS-1394

+++ tty/uart16550/uarthw-zynq.c
@@ -92,7 +92,7 @@ int uarthw_init(unsigned int uartn, void *hwctx, size_t hwctxsz, unsigned int *f
 {
 	void *base;
 	static const struct {
-		uint32_t base;
+		uintptr_t base;
 		unsigned int irqno;
 	} uarthw_info[] = {
 		{ .base = UART16550_BASE0, .irqno = UART16550_IRQ0 },
